% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{as.matrix.vrcfit}
\alias{as.matrix.vrcfit}
\title{Extract posterior draws as a matrix}
\usage{
\method{as.matrix}{vrcfit}(x, pars = NULL, ...)
}
\arguments{
\item{x}{A \code{vrcfit} object.}

\item{pars}{Optional character vector of parameter names.}

\item{...}{Passed to \code{rstan::as.matrix.stanfit}.}
}
\description{
Extract posterior draws as a matrix
}
