% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{beta}
\alias{beta}
\title{Beta prior}
\usage{
beta(shape1 = 1, shape2 = 1)
}
\arguments{
\item{shape1}{First shape parameter (alpha) (> 0).}

\item{shape2}{Second shape parameter (beta) (> 0).}
}
\value{
A prior specification list.
}
\description{
Beta prior
}
