% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{exponential}
\alias{exponential}
\title{Exponential prior}
\usage{
exponential(rate = 1)
}
\arguments{
\item{rate}{Rate parameter (> 0).}
}
\value{
A prior specification list.
}
\description{
Exponential prior
}
