% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{normal}
\alias{normal}
\title{Normal prior}
\usage{
normal(location = 0, scale = 1, autoscale = FALSE)
}
\arguments{
\item{location}{Prior mean.}

\item{scale}{Prior standard deviation (> 0).}

\item{autoscale}{Logical. If \code{TRUE}, the scale is adjusted by the column
scale of the design matrix (similar to rstanarm/epidemia). With the default
\code{standardise = TRUE}, autoscaling has no practical effect for continuous
predictors.}
}
\value{
A prior specification list.
}
\description{
Normal prior
}
