% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_mortality}
\alias{plot_mortality}
\title{Plot estimated mortality over time}
\usage{
plot_mortality(
  x,
  group_vars = c("region", "cause"),
  value = c("true_deaths", "rate"),
  ...
)
}
\arguments{
\item{x}{A \code{vrcfit} object.}

\item{group_vars}{Character vector of columns to group by in addition to \code{time}.
Defaults to \code{c("region", "cause")}.}

\item{value}{One of \code{"rate"} or \code{"true_deaths"}.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:theme]{ggplot2::theme()}}.}
}
\value{
A ggplot object.
}
\description{
Convenience plotting helper for visualising estimated latent mortality.
By default this plots expected true deaths (population x lambda), aggregated
over age and sex.
}
