% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{posterior_expected_counts}
\alias{posterior_expected_counts}
\title{Posterior summaries for expected observed counts}
\usage{
posterior_expected_counts(x, draws = FALSE, probs = c(0.1, 0.5, 0.9), ...)
}
\arguments{
\item{x}{A \code{vrcfit} object.}

\item{draws}{Logical. If TRUE, returns a matrix of posterior draws with one
column per observed cell.}

\item{probs}{Numeric vector of quantiles to compute.}

\item{...}{Reserved for future use.}
}
\description{
Extract posterior draws/summaries for the expected observed counts \code{mu}
returned from Stan as a generated quantity (\code{mu_rep}).
}
