% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{vrc_coef_summary}
\alias{vrc_coef_summary}
\title{Summarise covariate effects}
\usage{
vrc_coef_summary(x, probs = c(0.1, 0.5, 0.9), original_scale = FALSE)
}
\arguments{
\item{x}{A \code{vrcfit} object.}

\item{probs}{Quantiles to include.}

\item{original_scale}{Logical.}
}
\description{
Summarise covariate effects
}
