% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanmodels.R
\name{vrc_default_param_map}
\alias{vrc_default_param_map}
\title{Default mapping for core generated quantities}
\usage{
vrc_default_param_map()
}
\description{
Default mapping for core generated quantities
}
\keyword{internal}
