% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanmodels.R
\name{vrc_model}
\alias{vrc_model}
\title{Load and cache a Stan model}
\usage{
vrc_model(name = "vr_reporting_model", backend = c("rstan"), ...)
}
\arguments{
\item{name}{Stan model name (see \code{\link[=vrc_model_names]{vrc_model_names()}}) or a path to a \code{.stan}
file, or a \code{\link[=vrc_model_spec]{vrc_model_spec()}} object.}

\item{backend}{Backend to use. Currently only \code{"rstan"} is supported.}

\item{...}{Additional arguments passed to the backend model compiler.}
}
\value{
A compiled Stan model object.
}
\description{
Load and cache a Stan model
}
