% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanmodels.R
\name{vrc_model_names}
\alias{vrc_model_names}
\title{List Stan models shipped with the package}
\usage{
vrc_model_names()
}
\value{
Character vector of model names.
}
\description{
List Stan models shipped with the package
}
