% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{vrc_resolve_priors}
\alias{vrc_resolve_priors}
\title{Resolve priors into Stan hyperparameters}
\usage{
vrc_resolve_priors(priors, G, K_mort, K_rep, X_mort, X_rep)
}
\description{
Resolve priors into Stan hyperparameters
}
\keyword{internal}
