% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{vrc_sim_truth}
\alias{vrc_sim_truth}
\title{Extract cell-level truth from a simulated dataset}
\usage{
vrc_sim_truth(sim)
}
\arguments{
\item{sim}{A list returned by \code{\link[=vrc_simulate]{vrc_simulate()}}.}
}
\value{
A data.frame with additional columns:
\itemize{
\item \code{lambda_true} latent mortality rate
\item \code{rho_true} reporting completeness
\item \code{D_true} true deaths for the cause group
}
}
\description{
Helper for benchmarking. Given an object returned by \code{\link[=vrc_simulate]{vrc_simulate()}}, return
a data.frame aligned to \code{sim$df_full} containing cell-level true mortality
rates, reporting completeness, and true deaths.
}
