% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{vrcm}
\alias{vrcm}
\title{Fit a VR mortality + reporting model}
\usage{
vrcm(
  mortality = vrc_mortality(~1),
  reporting = vrc_reporting(~1),
  data,
  t0,
  use_mar_labels = FALSE,
  priors = NULL,
  ...
)
}
\arguments{
\item{mortality}{A \code{\link[=vrc_mortality]{vrc_mortality()}} specification.}

\item{reporting}{A \code{\link[=vrc_reporting]{vrc_reporting()}} specification.}

\item{data}{A data.frame in canonical VR long format.}

\item{t0}{Conflict start time.}

\item{use_mar_labels}{Logical.}

\item{priors}{Optional vrc_priors object.}

\item{...}{Passed to \code{\link[=vrc_fit]{vrc_fit()}}.}
}
\description{
Fit a VR mortality + reporting model
}
