% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{vrc_mortality}
\alias{vrc_mortality}
\title{Mortality model component specification}
\usage{
vrc_mortality(
  formula = ~1,
  facilities = NULL,
  conflict = c("fixed", "region"),
  time = c("national", "region")
)
}
\arguments{
\item{formula}{A model formula for additional mortality covariates.}

\item{facilities}{Optional character vector of facility level column names (perfect -> minimal).}

\item{conflict}{How to model conflict in mortality ("fixed" or "region").}

\item{time}{How to model time in mortality ("national" or "region").}
}
\description{
Create a mortality component specification for \code{\link[=vrcm]{vrcm()}}.
}
