% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{vrc_reporting}
\alias{vrc_reporting}
\title{Reporting model component specification}
\usage{
vrc_reporting(
  formula = ~1,
  facilities = NULL,
  conflict = c("fixed", "region"),
  time = c("national", "region")
)
}
\arguments{
\item{formula}{A model formula for additional reporting covariates.}

\item{facilities}{Optional character vector of facility level column names (perfect -> minimal).}

\item{conflict}{How to model conflict in reporting ("fixed" or "region").}

\item{time}{How to model time in reporting ("national" or "region").}
}
\description{
Create a reporting component specification for \code{\link[=vrcm]{vrcm()}}.
}
